refactor: isolate tool-output permission checks
